server/requirements.txt
# =============================================================================
# Core Dependencies (required)
# =============================================================================
torch>=2.0.0
torchaudio>=2.0.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Audio Processing
soundfile>=0.12.1
# Note: numpy version is managed by TTS/torch dependencies
librosa>=0.10.1
noisereduce>=3.0.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# =============================================================================
# TTS Backends (install the ones you need)
# =============================================================================

# XTTS v2 (Coqui TTS) - default backend
TTS>=0.22.0
# Pin transformers to avoid BeamSearchScorer import error (removed in transformers>=4.40)
transformers>=4.33.0,<4.40.0

# F5-TTS - high quality zero-shot voice cloning
f5-tts

# Chatterbox (Resemble AI) - production-ready zero-shot voice cloning
chatterbox-tts
